rename date-time-picker class names
